Have a single component for getting the user inputs that is placed in the two components where libs are created. Not being done the best way possible yet, but functions.
